pyproject.toml
[tool.poetry]
name = "next-gen"
version = "0.1.0"
description = ""
authors = ["ammar.elsherebasy <ammar.elsherebasy@fc-liefering.at>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.8"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
setuptools = "^80.9.0"
blinker = "^1.9.0"
click = "^8.2.1"
flask = "^3.1.1"
flask-cors = "^6.0.1"
flask-sqlalchemy = "^3.1.1"
itsdangerous = "^2.2.0"
jinja2 = "^3.1.6"
markupsafe = "^3.0.2"
sqlalchemy = "^2.0.41"
typing-extensions = "^4.14.1"
werkzeug = "^3.1.3"
dash = "^3.1.1"
google-cloud-bigquery = "^3.35.1"
numpy = "^2.3.2"
db-dtypes = "^1.4.3"
dash-bootstrap-components = "^2.0.3"
plotly = "^6.2.0"
rsconnect-python = "^1.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
